Extract functions into util files
