+++ client/src/utils/ui.ts
@@ -1,0 +1,6 @@
+export function cancelTextSelection() {
+  if (window.getSelection) {
+    const selection = window.getSelection();
+    if (selection) selection.removeAllRanges();
+  }
+}
